코드 리펙터링 및 적절한 http status 값 반환 및 해더 처리 (#29)
* add: TortoiseRouter.py 추가

Application.py 내부에 존재하던 클래스를 외부로 내보냄.

* up: db_config.json 파일 최신화

* del: 무분별하게 나뉘어진 Manager Class 삭제

* up: FileErrorCode 코드 추가 및 메소드 명을 더욱 명확하게 변경

* up: Static 하게 존재하던 다른 부분의 파일 Managing 함수들을 공통분모로 묶음

* up: Static 하게 존재하던 함수를 FileManger.py 로 이동 및 메타 클래스 삭제 메소드 명 간략화 등

* up: Information 조회 시 존재하지 않는 경우에 대해 예외 처리 추가

* up: 파일 업로드 성공시 헤더 정보에 Location 정보 추가 및 사용하지 않는 end-point 삭제

* up: 사용하지 않는 download end-point 삭제

* up: file delete 요청시 예외처리 추가

* up: File -> FileModel 명칭 업데이트

* up: 오타 수정

+++ Manager/File/FileManager.py
@@ -5,9 +5,10 @@ import uuid
 import aiofiles
 import aiofiles.os
 from sanic.request import File, Request
+from tortoise.exceptions import OperationalError
 
-from Common.UrlTool import get_query_string
-from Models.File import File as FileModel
+from Config.UploadConfig import UploadConfig
+from Models.FileModel import FileModel
 from .FileErrorCode import FileErrorCode
 
 
@@ -37,50 +38,51 @@ class FileManager:
         file_name, ext = os.path.splitext(file.name)
         return await FileModel.create(
             ext=ext,
-            mime_type=FileManager.guess_mime_type_by_file_name(file.name),
+            mime_type=mimetypes.guess_type(file.name)[0],
             file_size=len(file.body),
             user_id=user_id
         )
 
     @staticmethod
-    def guess_mime_type_by_file_name(file_name: str) -> str:
-        return mimetypes.guess_type(file_name)[0]
+    def get_abs_path_by_id(file_id: str) -> str:
+        """
+        Get abs path by id
+        :param file_id:
+        :return:
+        """
+        return f"{UploadConfig.get_instance().UPLOAD_ABS_PATH}/{file_id}"
 
     @staticmethod
     async def upload_file(file: File, file_id: str) -> None:
-        file_path = FileModel.get_file_path_by_id(file_id)
+        """
+        upload
+        :param file:
+        :param file_id:
+        :return:
+        """
+        file_path = FileManager.get_abs_path_by_id(file_id)
         async with aiofiles.open(file_path, 'wb') as fp:
             await fp.write(file.body)
         await fp.close()
 
     @staticmethod
-    async def upload_file_from_request(request: Request, file_key: str) -> tuple[FileErrorCode, str]:
-        user_id = request.form.get("userId")
-        error_code = FileManager.validate_files(request.files)
-        file_id: str = ""
-        if error_code == FileErrorCode.ERROR_SUCCESS:
-            file: File = request.files.get(file_key)
-            file_model = await FileManager.create_file_model(file, user_id)
-            file_id = str(file_model.pk)
-            request.app.add_task(FileManager.upload_file(file, file_id))
-        return error_code, file_id
+    async def delete_file_model(file_id: uuid, user_id: int) -> FileErrorCode:
+        """
 
-    @staticmethod
-    async def delete_file_from_request(request: Request):
-        query_string = get_query_string(request)
-        user_id = query_string.get("userId")
-        file_id = query_string.get("fileId")
-
-        await FileManager.delete_file_model(file_id, user_id)
-        await FileManager.delete_file(file_id)
-
-    @staticmethod
-    async def delete_file_model(file_id: uuid, user_id: int):
+        :param file_id:
+        :param user_id:
+        :return:
+        """
+        error_code = FileErrorCode.ERROR_SUCCESS
         model_to_delete = await FileModel.get(id=file_id, user_id=user_id)
-        await model_to_delete.delete()
+        if model_to_delete:
+            try:
+                await model_to_delete.delete()
+            except OperationalError:
+                error_code = FileErrorCode.ERROR_OPERATION
+
+        return error_code
 
     @staticmethod
     async def delete_file(file_id: uuid):
-        file_path = FileModel.get_file_path_by_id(file_id)
-
-        await aiofiles.os.remove(file_path)
+        await aiofiles.os.remove(FileManager.get_abs_path_by_id(file_id))
